\name{rrqr-package}
\alias{rrqr-package}
\alias{rrqr}
\docType{package}
\title{
\packageTitle{rrqr}
}
\description{
\packageDescription{rrqr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rrqr}
\packageIndices{rrqr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rrqr}

Maintainer: \packageMaintainer{rrqr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
