% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrqr.R
\name{rrqr}
\alias{rrqr}
\title{Add title here.}
\usage{
rrqr(X, tol = 1e-08)
}
\arguments{
\item{X}{Describe parameter here.}

\item{tol}{Describe parameter here.}
}
\description{
Add description here.
}
\examples{

set.seed(1)
X <- crossprod(matrix(rnorm(30),3,10))
rrqr(X)

}
